Calculate number of spares using system types #417

+++ inventory_management_system_api/services/item.py
@@ -3,28 +3,39 @@ Module for providing a service for managing items using the `ItemRepo`, `Catalog
 repositories.
 """
 
-from typing import Annotated, List, Optional
+import logging
+import random
+import time
+from contextlib import contextmanager
+from typing import Annotated, Generator, List, Optional
 
 from fastapi import Depends
+from pymongo.client_session import ClientSession
 
 from inventory_management_system_api.core.config import config
+from inventory_management_system_api.core.database import start_session_transaction
 from inventory_management_system_api.core.exceptions import (
     DatabaseIntegrityError,
     InvalidActionError,
     InvalidObjectIdError,
     MissingRecordError,
+    WriteConflictError,
 )
 from inventory_management_system_api.core.object_storage_api_client import ObjectStorageAPIClient
 from inventory_management_system_api.models.catalogue_item import PropertyOut
 from inventory_management_system_api.models.item import ItemIn, ItemOut
+from inventory_management_system_api.models.setting import SparesDefinitionOut
 from inventory_management_system_api.repositories.catalogue_category import CatalogueCategoryRepo
 from inventory_management_system_api.repositories.catalogue_item import CatalogueItemRepo
 from inventory_management_system_api.repositories.item import ItemRepo
+from inventory_management_system_api.repositories.setting import SettingRepo
 from inventory_management_system_api.repositories.system import SystemRepo
 from inventory_management_system_api.repositories.usage_status import UsageStatusRepo
 from inventory_management_system_api.schemas.catalogue_item import PropertyPostSchema
 from inventory_management_system_api.schemas.item import ItemPatchSchema, ItemPostSchema
 from inventory_management_system_api.services import utils
+
+logger = logging.getLogger()
 
 
 class ItemService:
@@ -41,6 +52,7 @@ class ItemService:
         catalogue_item_repository: Annotated[CatalogueItemRepo, Depends(CatalogueItemRepo)],
         system_repository: Annotated[SystemRepo, Depends(SystemRepo)],
         usage_status_repository: Annotated[UsageStatusRepo, Depends(UsageStatusRepo)],
+        setting_repository: Annotated[SettingRepo, Depends(SettingRepo)],
     ) -> None:
         """
         Initialise the `ItemService` with an `ItemRepo`, `CatalogueCategoryRepo`,
@@ -51,12 +63,14 @@ class ItemService:
         :param catalogue_item_repository: The `CatalogueItemRepo` repository to use.
         :param system_repository: The `SystemRepo` repository to use.
         :param usage_status_repository: The `UsageStatusRepo` repository to use.
+        :param setting_repository: The `SettingRepo` repository to use.
         """
         self._item_repository = item_repository
         self._catalogue_category_repository = catalogue_category_repository
         self._catalogue_item_repository = catalogue_item_repository
         self._system_repository = system_repository
         self._usage_status_repository = usage_status_repository
+        self._setting_repository = setting_repository
 
     def create(self, item: ItemPostSchema) -> ItemOut:
         """
@@ -93,9 +107,14 @@ class ItemService:
         defined_properties = catalogue_category.properties
         properties = utils.process_properties(defined_properties, supplied_properties)
 
-        return self._item_repository.create(
-            ItemIn(**{**item.model_dump(), "properties": properties, "usage_status": usage_status.value})
-        )
+        # Update number of spares when creating
+        with self._start_transaction_impacting_number_of_spares(
+            "creating item", catalogue_item_id, item.system_id
+        ) as session:
+            return self._item_repository.create(
+                ItemIn(**{**item.model_dump(), "properties": properties, "usage_status": usage_status.value}),
+                session=session,
+            )
 
     def get(self, item_id: str) -> Optional[ItemOut]:
         """
@@ -116,6 +135,7 @@ class ItemService:
         """
         return self._item_repository.list(system_id, catalogue_item_id)
 
+    # pylint:disable=too-many-locals
     def update(self, item_id: str, item: ItemPatchSchema) -> ItemOut:
         """
         Update an item by its ID.
@@ -139,7 +159,8 @@ class ItemService:
         if "catalogue_item_id" in update_data and item.catalogue_item_id != stored_item.catalogue_item_id:
             raise InvalidActionError("Cannot change the catalogue item the item belongs to")
 
-        if "system_id" in update_data and item.system_id != stored_item.system_id:
+        moving_system = "system_id" in update_data and item.system_id != stored_item.system_id
+        if moving_system:
             system = self._system_repository.get(item.system_id)
             if not system:
                 raise MissingRecordError(f"No system found with ID: {item.system_id}")
@@ -172,7 +193,21 @@ class ItemService:
 
             update_data["properties"] = utils.process_properties(defined_properties, supplied_properties)
 
+        # Moving system could effect the number of spares of the catalogue item as the type of the system might be
+        # different
+        if moving_system:
+            # Can't currently move items, so we can just write lock the stored catalogue item as opposed to checking
+            # the update data.
+            with self._start_transaction_impacting_number_of_spares(
+                "updating item", stored_item.catalogue_item_id, dest_system_id=item.system_id
+            ) as session:
+                return self._item_repository.update(
+                    item_id, ItemIn(**{**stored_item.model_dump(), **update_data}), session=session
+                )
+
         return self._item_repository.update(item_id, ItemIn(**{**stored_item.model_dump(), **update_data}))
+
+    # pylint:enable=too-many-locals
 
     def delete(self, item_id: str, access_token: Optional[str] = None) -> None:
         """
@@ -180,13 +215,20 @@ class ItemService:
 
         :param item_id: The ID of the item to delete.
         :param access_token: The JWT access token to use for auth with the Object Storage API if object storage enabled.
+        :raises MissingRecordError: If the item doesn't exist.
         """
+        item = self.get(item_id)
+        if item is None:
+            raise MissingRecordError(f"No item found with ID: {item_id}")
+
         # First, attempt to delete any attachments and/or images that might be associated with this item.
         if config.object_storage.enabled:
             ObjectStorageAPIClient.delete_attachments(item_id, access_token)
             ObjectStorageAPIClient.delete_images(item_id, access_token)
 
-        return self._item_repository.delete(item_id)
+        # Deleting could effect the number of spares of the catalogue item if this one is currently a spare
+        with self._start_transaction_impacting_number_of_spares("deleting item", item.catalogue_item_id) as session:
+            return self._item_repository.delete(item_id, session=session)
 
     def _merge_missing_properties(
         self, properties: List[PropertyOut], supplied_properties: List[PropertyPostSchema]
@@ -213,3 +255,78 @@ class ItemService:
             else:
                 merged_properties.append(PropertyPostSchema(**prop.model_dump()))
         return merged_properties
+
+    @contextmanager
+    def _start_transaction_impacting_number_of_spares(
+        self, action_description: str, catalogue_item_id: str, dest_system_id: Optional[str] = None
+    ) -> Generator[Optional[ClientSession], None, None]:
+        """
+        Handles recalculation of the `number_of_spares` field of a catalogue item for updates that will impact it but
+        only when there is a spares definition is set.
+
+        When necessary, starts a MongoDB session and transaction, then write locks the catalogue item before yielding to
+        allow an update to take place using the returned session. Once any tasks using the session context have finished
+        it will finish by recalculating the number of spares for the catalogue item before finishing the transaction.
+        This write lock prevents similar actions from occurring during the update to prevent an incorrect update e.g. if
+        another item was added between counting the documents and then updating the `number_of_spares` field it would
+        cause a miscount. It also ensures any action executed using the session will either fail or succeed with the
+        spares update.
+
+        :param action_description: Description of what the contents of the transaction is doing so it can be used in
+                                   any logging or raise errors.
+        :param catalogue_item_id: ID of the effected catalogue item which will need its `number_of_spares` field
+                                  updating.
+        :param dest_system_id: ID of the system being put in/moved to (if applicable). Will be write locked to prevent
+                               editing of system type after counting spares to avoid miscounts.
+        """
+
+        # Firstly obtain the spares definition to figure out if it is defined or not
+        spares_definition = self._setting_repository.get(SparesDefinitionOut)
+
+        if spares_definition is None:
+            # No session/transaction is needed as there is no spares update to perform
+            yield None
+        else:
+            # Particularly when creating multiple items within the same catalogue item in quick succession, multiple
+            # conflicting requests can occur. To reduce the chances we retry such requests so that the default 5ms
+            # transaction timeout is less of an issue.
+            start_time = time.time()
+            retry = True
+            while retry:
+                try:
+                    with start_session_transaction(action_description) as session:
+                        # Write lock the catalogue item to prevent any other updates from occurring during the rest of
+                        # the transaction
+                        self._catalogue_item_repository.update_number_of_spares(
+                            catalogue_item_id, None, session=session
+                        )
+
+                        # Write lock the destination system
+                        # This will prevent the case where a system has no items currently, allowing the system type to
+                        # be modified after the count but before the update finishes and instead force conflicts with
+                        # system type modifications.
+                        if dest_system_id:
+                            self._system_repository.write_lock(dest_system_id, session)
+
+                        # Allow any other updates to occur using the same session
+                        yield session
+
+                        # Obtain and update the number of spares
+                        logger.info("Updating the number of spares of the catalogue item with ID %s", catalogue_item_id)
+                        number_of_spares = self._item_repository.count_in_catalogue_item_with_system_type_one_of(
+                            catalogue_item_id,
+                            [system_type.id for system_type in spares_definition.system_types],
+                            session=session,
+                        )
+                        self._catalogue_item_repository.update_number_of_spares(
+                            catalogue_item_id, number_of_spares, session=session
+                        )
+                        retry = False
+                except WriteConflictError as exc:
+                    # Keep retrying, but only if we have been retrying for less than 5 seconds so we dont let the
+                    # request take too long and leave potential for it to block other requests if the threadpool is full
+                    if time.time() - start_time > 5:
+                        raise exc
+                    # Wait some random time as there is no point in retrying immediately if we are already write
+                    # locked. Between 10ms and 50ms.
+                    time.sleep(random.uniform(0.01, 0.05))
